+++ src/main/java/com/finnect/user/application/AuthorizeService.java
@@ -5,12 +5,14 @@ import com.finnect.user.application.port.in.command.AuthorizeCommand;
 import com.finnect.user.application.port.out.ObtainAuthenticationPort;
 import com.finnect.user.domain.AccessToken;
 import com.finnect.user.domain.state.UserAuthenticationState;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class AuthorizeService implements AuthorizeUseCase {
 
     private final ObtainAuthenticationPort obtainAuthenticationPort;
