+++ src/main/java/com/finnect/user/application/FindUsernameService.java
@@ -9,12 +9,14 @@ import com.finnect.user.application.port.out.LoadUserPort;
 import com.finnect.user.application.port.out.SaveEmailCodePort;
 import com.finnect.user.domain.EmailCode;
 import com.finnect.user.domain.User;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class FindUsernameService implements FindUsernameUseCase {
 
     private final LoadUserPort loadUserPort;
